initial commit
Basic student class

+++ Student.java
@@ -1,4 +1,4 @@
 public class Student extends Person
 {
-  float gpa;
+   double gpa;
 }
